revert and improve

+++ ct/fladder.sh
@@ -24,6 +24,11 @@ function update_script() {
   check_container_storage
   check_container_resources
 
+  if [[ ! -d /opt/fladder ]]; then
+    msg_error "No ${APP} Installation Found!"
+    exit
+  fi
+
   if check_for_gh_release "Fladder" "DonutWare/Fladder"; then
     msg_info "Stopping Service"
     systemctl stop nginx
